Move things around in subspace-node to not expose unnecessary things via library API
