--- crates/subspace-node/src/cli.rs
@@ -1,72 +1,0 @@
-// Copyright (C) 2021 Subspace Labs, Inc.
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-//! Subspace CLI library.
-
-use clap::Parser;
-
-/// Utilities for working with a node.
-#[derive(Debug, clap::Subcommand)]
-pub enum Subcommand {
-    /// Key management cli utilities
-    #[clap(subcommand)]
-    Key(sc_cli::KeySubcommand),
-
-    /// Build a chain specification.
-    BuildSpec(sc_cli::BuildSpecCmd),
-
-    /// Validate blocks.
-    CheckBlock(sc_cli::CheckBlockCmd),
-
-    /// Export blocks.
-    ExportBlocks(sc_cli::ExportBlocksCmd),
-
-    /// Export the state of a given block into a chain spec.
-    ExportState(sc_cli::ExportStateCmd),
-
-    /// Import blocks.
-    ImportBlocks(sc_cli::ImportBlocksCmd),
-
-    /// Remove the whole chain.
-    PurgeChain(sc_cli::PurgeChainCmd),
-
-    /// Revert the chain to a previous state.
-    Revert(sc_cli::RevertCmd),
-
-    /// The custom benchmark subcommand benchmarking runtime pallets.
-    #[clap(name = "benchmark", about = "Benchmark runtime pallets.")]
-    Benchmark(frame_benchmarking_cli::BenchmarkCmd),
-}
-
-/// Command used to run a Subspace node.
-#[derive(Debug, Parser)]
-pub struct RunCmd {
-    /// Base command to run a node.
-    #[clap(flatten)]
-    pub base: sc_cli::RunCmd,
-}
-
-/// Subspace Cli.
-#[derive(Debug, Parser)]
-pub struct Cli {
-    /// Various utility commands.
-    #[clap(subcommand)]
-    pub subcommand: Option<Subcommand>,
-
-    /// Run a node.
-    #[clap(flatten)]
-    pub run: RunCmd,
-}
