--- crates/subspace-node/src/command.rs
@@ -1,254 +1,0 @@
-// Copyright (C) 2021 Subspace Labs, Inc.
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-use crate::chain_spec;
-use crate::cli::{Cli, Subcommand};
-use futures::future::TryFutureExt;
-use sc_cli::{ChainSpec, RuntimeVersion, SubstrateCli};
-use sc_executor::NativeExecutionDispatch;
-use sp_core::crypto::Ss58AddressFormat;
-use subspace_runtime::RuntimeApi;
-
-/// Subspace node error.
-#[derive(thiserror::Error, Debug)]
-pub enum Error {
-    /// Subspace service error.
-    #[error(transparent)]
-    SubspaceService(#[from] subspace_service::Error),
-
-    /// CLI error.
-    #[error(transparent)]
-    SubstrateCli(#[from] sc_cli::Error),
-
-    /// Substrate service error.
-    #[error(transparent)]
-    SubstrateService(#[from] sc_service::Error),
-
-    /// Other kind of error.
-    #[error("Other: {0}")]
-    Other(String),
-}
-
-impl From<String> for Error {
-    fn from(s: String) -> Self {
-        Self::Other(s)
-    }
-}
-
-struct ExecutorDispatch;
-
-impl NativeExecutionDispatch for ExecutorDispatch {
-    /// Only enable the benchmarking host functions when we actually want to benchmark.
-    #[cfg(feature = "runtime-benchmarks")]
-    type ExtendHostFunctions = frame_benchmarking::benchmarking::HostFunctions;
-    /// Otherwise we only use the default Substrate host functions.
-    #[cfg(not(feature = "runtime-benchmarks"))]
-    type ExtendHostFunctions = ();
-
-    fn dispatch(method: &str, data: &[u8]) -> Option<Vec<u8>> {
-        subspace_runtime::api::dispatch(method, data)
-    }
-
-    fn native_version() -> sc_executor::NativeVersion {
-        subspace_runtime::native_version()
-    }
-}
-
-impl SubstrateCli for Cli {
-    fn impl_name() -> String {
-        "Subspace".into()
-    }
-
-    fn impl_version() -> String {
-        env!("SUBSTRATE_CLI_IMPL_VERSION").into()
-    }
-
-    fn description() -> String {
-        env!("CARGO_PKG_DESCRIPTION").into()
-    }
-
-    fn author() -> String {
-        env!("CARGO_PKG_AUTHORS").into()
-    }
-
-    fn support_url() -> String {
-        "https://discord.gg/vhKF9w3x".into()
-    }
-
-    fn copyright_start_year() -> i32 {
-        2021
-    }
-
-    fn load_spec(&self, id: &str) -> Result<Box<dyn sc_service::ChainSpec>, String> {
-        Ok(match id {
-            "testnet" => Box::new(chain_spec::subspace_testnet_config()?),
-            "dev" => Box::new(chain_spec::subspace_development_config()?),
-            "" | "local" => Box::new(chain_spec::subspace_local_testnet_config()?),
-            path => Box::new(chain_spec::SubspaceChainSpec::from_json_file(
-                std::path::PathBuf::from(path),
-            )?),
-        })
-    }
-
-    fn native_runtime_version(_: &Box<dyn ChainSpec>) -> &'static RuntimeVersion {
-        &subspace_runtime::VERSION
-    }
-}
-
-fn set_default_ss58_version<C: AsRef<dyn ChainSpec>>(chain_spec: C) {
-    let maybe_ss58_address_format = chain_spec
-        .as_ref()
-        .properties()
-        .get("ss58Format")
-        .map(|v| {
-            v.as_u64()
-                .expect("ss58Format must always be an unsigned number; qed")
-        })
-        .map(|v| {
-            v.try_into()
-                .expect("ss58Format must always be within u16 range; qed")
-        })
-        .map(Ss58AddressFormat::custom);
-
-    if let Some(ss58_address_format) = maybe_ss58_address_format {
-        sp_core::crypto::set_default_ss58_version(ss58_address_format);
-    }
-}
-
-/// Parse and run command line arguments
-pub fn run() -> std::result::Result<(), Error> {
-    let cli = Cli::from_args();
-
-    match &cli.subcommand {
-        Some(Subcommand::Key(cmd)) => cmd.run(&cli)?,
-        Some(Subcommand::BuildSpec(cmd)) => {
-            let runner = cli.create_runner(cmd)?;
-            runner.sync_run(|config| cmd.run(config.chain_spec, config.network))?
-        }
-        Some(Subcommand::CheckBlock(cmd)) => {
-            let runner = cli.create_runner(cmd)?;
-            set_default_ss58_version(&runner.config().chain_spec);
-            runner.async_run(|config| {
-                let sc_service::PartialComponents {
-                    client,
-                    import_queue,
-                    task_manager,
-                    ..
-                } = subspace_service::new_partial::<RuntimeApi, ExecutorDispatch>(&config)?;
-                Ok((
-                    cmd.run(client, import_queue).map_err(Error::SubstrateCli),
-                    task_manager,
-                ))
-            })?;
-        }
-        Some(Subcommand::ExportBlocks(cmd)) => {
-            let runner = cli.create_runner(cmd)?;
-            set_default_ss58_version(&runner.config().chain_spec);
-            runner.async_run(|config| {
-                let sc_service::PartialComponents {
-                    client,
-                    task_manager,
-                    ..
-                } = subspace_service::new_partial::<RuntimeApi, ExecutorDispatch>(&config)?;
-                Ok((
-                    cmd.run(client, config.database)
-                        .map_err(Error::SubstrateCli),
-                    task_manager,
-                ))
-            })?;
-        }
-        Some(Subcommand::ExportState(cmd)) => {
-            let runner = cli.create_runner(cmd)?;
-            set_default_ss58_version(&runner.config().chain_spec);
-            runner.async_run(|config| {
-                let sc_service::PartialComponents {
-                    client,
-                    task_manager,
-                    ..
-                } = subspace_service::new_partial::<RuntimeApi, ExecutorDispatch>(&config)?;
-                Ok((
-                    cmd.run(client, config.chain_spec)
-                        .map_err(Error::SubstrateCli),
-                    task_manager,
-                ))
-            })?;
-        }
-        Some(Subcommand::ImportBlocks(cmd)) => {
-            let runner = cli.create_runner(cmd)?;
-            set_default_ss58_version(&runner.config().chain_spec);
-            runner.async_run(|config| {
-                let sc_service::PartialComponents {
-                    client,
-                    import_queue,
-                    task_manager,
-                    ..
-                } = subspace_service::new_partial::<RuntimeApi, ExecutorDispatch>(&config)?;
-                Ok((
-                    cmd.run(client, import_queue).map_err(Error::SubstrateCli),
-                    task_manager,
-                ))
-            })?;
-        }
-        Some(Subcommand::PurgeChain(cmd)) => {
-            let runner = cli.create_runner(cmd)?;
-            runner.sync_run(|config| cmd.run(config.database))?
-        }
-        Some(Subcommand::Revert(cmd)) => {
-            let runner = cli.create_runner(cmd)?;
-            set_default_ss58_version(&runner.config().chain_spec);
-            runner.async_run(|config| {
-                let sc_service::PartialComponents {
-                    client,
-                    backend,
-                    task_manager,
-                    ..
-                } = subspace_service::new_partial::<RuntimeApi, ExecutorDispatch>(&config)?;
-                Ok((
-                    cmd.run(client, backend).map_err(Error::SubstrateCli),
-                    task_manager,
-                ))
-            })?;
-        }
-        Some(Subcommand::Benchmark(cmd)) => {
-            if cfg!(feature = "runtime-benchmarks") {
-                let runner = cli.create_runner(cmd)?;
-                set_default_ss58_version(&runner.config().chain_spec);
-                runner.sync_run(|config| {
-                    cmd.run::<subspace_runtime::Block, ExecutorDispatch>(config)
-                })?;
-            } else {
-                return Err(Error::Other(
-                    "Benchmarking wasn't enabled when building the node. You can enable it with \
-                    `--features runtime-benchmarks`."
-                        .into(),
-                ));
-            }
-        }
-        None => {
-            let runner = cli.create_runner(&cli.run.base)?;
-            set_default_ss58_version(&runner.config().chain_spec);
-            runner.run_node_until_exit(|config| async move {
-                subspace_service::new_full::<subspace_runtime::RuntimeApi, ExecutorDispatch>(
-                    config, true,
-                )
-                .await
-                .map(|full| full.task_manager)
-            })?;
-        }
-    }
-
-    Ok(())
-}
