+++ crates/subspace-node/src/lib.rs
@@ -16,10 +16,103 @@
 
 //! Subspace Node library.
 
-#![warn(missing_docs)]
-
 mod chain_spec;
-pub mod cli;
-mod command;
 
-pub use self::command::{run, Error};
+use clap::Parser;
+use sc_cli::SubstrateCli;
+use sc_executor::RuntimeVersion;
+use sc_service::ChainSpec;
+
+/// Utilities for working with a node.
+#[derive(Debug, clap::Subcommand)]
+pub enum Subcommand {
+    /// Key management cli utilities
+    #[clap(subcommand)]
+    Key(sc_cli::KeySubcommand),
+
+    /// Build a chain specification.
+    BuildSpec(sc_cli::BuildSpecCmd),
+
+    /// Validate blocks.
+    CheckBlock(sc_cli::CheckBlockCmd),
+
+    /// Export blocks.
+    ExportBlocks(sc_cli::ExportBlocksCmd),
+
+    /// Export the state of a given block into a chain spec.
+    ExportState(sc_cli::ExportStateCmd),
+
+    /// Import blocks.
+    ImportBlocks(sc_cli::ImportBlocksCmd),
+
+    /// Remove the whole chain.
+    PurgeChain(sc_cli::PurgeChainCmd),
+
+    /// Revert the chain to a previous state.
+    Revert(sc_cli::RevertCmd),
+
+    /// The custom benchmark subcommand benchmarking runtime pallets.
+    #[clap(name = "benchmark", about = "Benchmark runtime pallets.")]
+    Benchmark(frame_benchmarking_cli::BenchmarkCmd),
+}
+
+/// Command used to run a Subspace node.
+#[derive(Debug, Parser)]
+pub struct RunCmd {
+    /// Base command to run a node.
+    #[clap(flatten)]
+    pub base: sc_cli::RunCmd,
+}
+
+/// Subspace Cli.
+#[derive(Debug, Parser)]
+pub struct Cli {
+    /// Various utility commands.
+    #[clap(subcommand)]
+    pub subcommand: Option<Subcommand>,
+
+    /// Run a node.
+    #[clap(flatten)]
+    pub run: RunCmd,
+}
+
+impl SubstrateCli for Cli {
+    fn impl_name() -> String {
+        "Subspace".into()
+    }
+
+    fn impl_version() -> String {
+        env!("SUBSTRATE_CLI_IMPL_VERSION").into()
+    }
+
+    fn description() -> String {
+        env!("CARGO_PKG_DESCRIPTION").into()
+    }
+
+    fn author() -> String {
+        env!("CARGO_PKG_AUTHORS").into()
+    }
+
+    fn support_url() -> String {
+        "https://discord.gg/vhKF9w3x".into()
+    }
+
+    fn copyright_start_year() -> i32 {
+        2021
+    }
+
+    fn load_spec(&self, id: &str) -> Result<Box<dyn ChainSpec>, String> {
+        Ok(match id {
+            "testnet" => Box::new(chain_spec::subspace_testnet_config()?),
+            "dev" => Box::new(chain_spec::subspace_development_config()?),
+            "" | "local" => Box::new(chain_spec::subspace_local_testnet_config()?),
+            path => Box::new(chain_spec::SubspaceChainSpec::from_json_file(
+                std::path::PathBuf::from(path),
+            )?),
+        })
+    }
+
+    fn native_runtime_version(_: &Box<dyn ChainSpec>) -> &'static RuntimeVersion {
+        &subspace_runtime::VERSION
+    }
+}
