+++ cumulus/parachain-template/node/src/cli.rs
@@ -92,7 +92,7 @@ pub struct Cli {
 #[derive(Debug)]
 pub struct RelayChainCli {
 	/// The actual relay chain cli object.
-	pub base: subspace_node::cli::RunCmd,
+	pub base: subspace_node::RunCmd,
 
 	/// Optional chain id that should be passed to the relay chain.
 	pub chain_id: Option<String>,
@@ -111,6 +111,6 @@ impl RelayChainCli {
 		let chain_id = extension.map(|e| e.relay_chain.clone());
 		let base_path = para_config.base_path.as_ref().map(|x| x.path().join("polkadot"));
 		// TODO: we might want to forcibly inject the `--validator` flag.
-		Self { base_path, chain_id, base: subspace_node::cli::RunCmd::parse_from(relay_chain_args) }
+		Self { base_path, chain_id, base: subspace_node::RunCmd::parse_from(relay_chain_args) }
 	}
 }
